feat: 아티클 상세 조회 요청 메서드 및 타입 구현
