+++ models/author.ts
@@ -1,0 +1,6 @@
+export type Author = {
+  username: string;
+  bio: string;
+  image: string;
+  following: boolean;
+};
